Revert rename of existing classes
