+++ sql/catalyst/src/main/java/org/apache/spark/sql/catalyst/util/Geography.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.catalyst.util.geo;
+package org.apache.spark.sql.catalyst.util;
 
 import org.apache.spark.unsafe.types.GeographyVal;
