+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/util/GeometryExecutionSuite.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.catalyst.util.geo;
+package org.apache.spark.sql.catalyst.util;
 
 import org.apache.spark.unsafe.types.GeometryVal;
 import org.junit.jupiter.api.Test;
